Repo setup & viewer version 7.1.13.53431 ebuild
unpacker is needed for uncompressing zst files.
Downloading 3rd-party files as part of the upstream CMake configuration
wasn't allowed, the files need to be registered as part of the emerge
process, but the unpacking destination path didn't seem customisable, so
the hack is to just move the unpacked files to where they are needed to
be in upstream building, rather than patching upstream files to work
with Portage's default.
The keyword is not masked because 7.1.13 is expected to be the default
(stable) version soon.. even though the Gentoo port still crashes
sometimes (might just be leftover cache from wrong CMake configuration
build type, explained next).
This ebuild is not a live Git using one, so the upstream configuration
process can't figure out the revision number, so it needs to be set on
the environment.
Setting the CMake build type in mycmakeargs didn't have any effect,
emerge set it to RelWithDebInfo instead (maybe that's why it would often
crash, and the way it's supposed to be done is to set it to the cmake
eclass CMAKE_BUILD_TYPE variable instead.
Patching Collada DOM also seemed to have to be done as part of the
emerge process (the one in the upstream configuration process didn't
seem to have any effect).
No USEs yet for now, and the WebRTC build being used is Crow Misia's for
now, as it's been proven to be the safer bet not to prevent CEF from
working.