7.1.13 has been upgraded to revision 53452
The optimisation flag environment custom setting is removed so that it
won't override user's in /etc/portage/make.conf.
The C++ standard environment custom setting is removed because Portage
already has a default for this ebuild, which is gnu++20, that I don't
think the custom one would cause anything significantly different, if
there's any.