Megapahit viewer 7.2.0-r54105 ebuild
Introducing the new discord (for Discord Social SDK) USE flag, which is
off by default.
I don't know how to pass the value of Portage's ${DISTDIR} dynamically
to the patched file yet, so I hardcoded it with `/var/cache/distfiles`,
which should be the most common at the time of this writing.